packages: set file limits for all services
Of all the DC/OS services, only spartan and mesos-{master,slave,public-slave}
set file limits. In all other cases the limit defaults to whatever is the default
for the distribution. In the case of CoreOS the default soft limit is 1024.

This PR sets the default file limit of all other dcos-* services to 16384.

Ideally this kind of change should inspect a very large, very busy cluster
and set the limits accordingly. The closest thing we have at this stage
is the soak cluster. The introduced limit of 16384 is far higher than any
services on the soak cluster are currently allowed to allocate so this shouldn't
constrain any existing service, given the information available to us at this stage.

In future, as our monitoring improves, we will collect peak usage from large
clusters and adjust the limits accordingly.

Ideally all services should run with a predictable maximum number of open files
to support far tighter limits.

Fixes https://mesosphere.atlassian.net/browse/DCOS-13448