Adminrouter now uses adminrouter-listen-ee.conf
This change is justified by the fact that we need to have a file
that has a same path (adminrouter-listen.conf) for both EE and Open,
but different contents, both when it comes to master/agent node type
and the EE/Open  DC/OS flavour. Currently genconf only support
per-node-type differentiation.

The differentiation stems from adminrouter listening on all interfaces
by default. This break test harness as it tries to emulate agent AR
for some of the tests. So the idea is to provide different
adminrouter-listen.conf file for tests - a one that makes AR listen
only on 127.0.0.1.