Testing Mesos v1 agent operator API authn/z
*Note:* This PR replaces #409

The new debugging API uses the agent v1 operator API, but all v1 APIs
were previously blacklisted in strict mode for Argus, for the lack of
consistent authn/z. In order to remove the agent v1 operator API from
the security blacklist, we must test the authn/z of each Call in the
API. This PR adds those tests.

[DCOS-11255](https://mesosphere.atlassian.net/browse/DCOS-11255)
[DCOS-9574](https://mesosphere.atlassian.net/browse/DCOS-9574)

 - [x] Included a test which will fail if code is reverted but test is not
 - [x] Followed relevant code rules [Rules for Packages and Systemd](https://github.com/dcos/dcos/tree/master/docs)