Changed '/etc_master/dcos-backup-master.json' to a YAML file
Previously, tests were breaking in 'disabled' mode because
'ssl_enabled == false' in this mode, causing a trailing comma to be
present in the resulting JSON. Rather than add more conditionals or
rearrange the configuration file to accommodate this, I decided to
update the configuration to YAML (the preferred configuration format
going forward). This format doesn't use commas to separate fields, so it
doesn't have this issue any longer.