curl: adjust --with-ca-bundle path to new requests/certifi structure
Python's requests package was updates from 2.10 to 2.18 in a previous
commit. The new Python requests package does not bring along its own
cacerts.pem (a collection of trust anchors) file anymore, but instead
uses certifi which evolved as an independent project for distributing
the Mozilla CA bundle to Python applications.